Reduce idle cpu usage in watch mode
This PR reverts back to use chokidar for watching instead of esbuild's
built-in watch functionality. The watch mode in esbuild is based on
polling and hovers around 50% - 60% CPU usage on my
Macbook Air M1 2020 while idling.
Chokidar on the other hand uses the OS'es filesystem events and only
consumes around 10% CPU usage on the same machine