Do not respond to sourcemaps for unknown files
This is difficult to test for, but when using dev tools to inspect mocha tests, you'll eventually requests sourcemaps. One of those sourcemaps is for `node_modules/mocha/mocha.js.map`. Importantly, `mocha` is a framework, is injected as a separate part of the karma process, and is not bundled by esbuild.

The old sourcemap middleware would catch this request, then wait on esbuild to finally build `mocha.js`. But that's not going to happen, so we end up permanently waiting.