Add formatError to unminify stack traces
While working on module mode, I've realized that it's just impossible to provide individual file modules and hook into Karma's crapy served+preprocessed `file.sourceMap` requirement.

So, this hooks into the `config.formatError` hook that the error reporter uses, and performs the same actions. But now we can populate this formatter with any number of bundles that we're generating.