Prevent concurrent builds from happening
AMP noticed that multiple _initial_ builds can happen if Karma takes a long time between calls to karma-esbuild's preprocessor (in our case, we have hundreds of files having SHA hashes performed). This lead to multiple `bundle.write` calls happening.

Now, our test suite is so large, esbuild takes a few seconds to run through everything. So when the second `bundle.write` happens, we're still trying to bundle the first time. This leads to 2 initial builds. Because both builds are `incremental`, they hold on to esbuild's Go process. And because we would lose our reference to the first incremental build when the second finished, we now had an undying holder on esbuild.

While solving this concurrent initial build problem, I decided that we should never have 2 concurrent builds happening at the same time. Now only 1 build can be happening at a time, and if multiple write come in (either during the initial build or during a rebuild), we'll wait for the in-progress build to finish before starting the new one.

Async reentrant problems are hard.